Move the code to avoid cycle dependency
The xmlManager can't be placed in tmf core ui, because it depends on the
optional xml analysis plug-in. A cycle in the dependencies causes a failure in
the build. This failure can be disabled in Eclipse, but the batch maven build
is also affected, such that this error must be fixed.

Just move the xmlManager into as XML analysis view.

TODO:
  - Handle file delete (see XmlUtils.java)
  - Check the icon

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>